Add min_guaranteed_pool_size, refactor coordinator integration
New pool-level config field min_guaranteed_pool_size: minimum connections
per user protected from coordinator eviction, independent of min_pool_size
(which controls prewarm/replenish). Effective protection for eviction is
max(user.min_pool_size, pool.min_guaranteed_pool_size). Solves the common
case where min_pool_size=0 (no prewarm) but eviction fairness is needed.

Refactoring:
- Check reserve_tx.send() failure instead of silent drop
- Snapshot spare_above_min() once per eviction candidate (avoids TOCTOU)
- Replace unwrap() with expect() on ReserveGrant consumption
- Use Arc::clone() for explicit reference counting
- Compare connection age in u128 space (no truncation)
- Consistent log prefix [pool: X][user: Y] for eviction messages
- ConnectionPool.coordinator visibility narrowed to pub(crate)
- Remove stale stub documentation

+++ src/pool/inner.rs
@@ -660,7 +660,7 @@ impl Pool {
     /// Returns `true` if a connection was evicted.
     pub fn evict_one_idle(&self, min_lifetime_ms: u64) -> bool {
         self.retain_oldest_first(
-            |_, metrics| metrics.age().as_millis() as u64 >= min_lifetime_ms,
+            |_, metrics| metrics.age().as_millis() >= u128::from(min_lifetime_ms),
             1,
         ) > 0
     }
